chore: Add script for running semgrep on contracts
This commit adds a new script `semgrep.sh` that allows running semgrep on contracts. The script takes a contracts folder as an argument and performs semgrep analysis on all `.sol` files in parallel batches. The results are saved in a specified directory. This script improves the efficiency of running semgrep on contracts.